split out pre-onboarding into it's own queue
